# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactKind": {
        "enum": [
          "TEXT",
          "TABLE",
          "CHART",
          "SQL",
          "CODE",
          "DIFF",
          "REPORT",
          "DASHBOARD",
          "FILE",
          "DIAGRAM"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
      "ArtifactView": {
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inline_content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inline Content"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "lineage": {
            "additionalProperties": true,
            "title": "Lineage",
            "type": "object"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Key"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "run_id",
          "kind",
          "title",
          "media_type",
          "inline_content",
          "storage_key",
          "classification",
          "lineage",
          "created_at"
        ],
        "title": "ArtifactView",
        "type": "object"
      },
...
      "ClaimView": {
        "properties": {
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "critic_notes": {
            "additionalProperties": true,
            "title": "Critic Notes",
            "type": "object"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "ordinal",
          "statement",
          "confidence",
          "verification_status",
          "critic_notes"
        ],
        "title": "ClaimView",
        "type": "object"
      },
      "Classification": {
        "enum": [
          "PUBLIC",
          "INTERNAL",
          "CONFIDENTIAL",
          "RESTRICTED"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "EvidenceView": {
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingested_at": {
            "format": "date-time",
            "title": "Ingested At",
            "type": "string"
          },
          "lineage": {
            "additionalProperties": true,
            "title": "Lineage",
            "type": "object"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "step_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          }
        },
        "required": [
          "id",
          "run_id",
          "step_id",
          "evidence_type",
          "source",
          "resource",
          "observed_at",
          "ingested_at",
          "content",
          "confidence",
          "classification",
          "lineage"
        ],
        "title": "EvidenceView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunStepView": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "depends_on": {
            "items": {
              "type": "integer"
            },
            "title": "Depends On",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/StepKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "output_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Ref"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/StepStatus"
          }
        },
        "required": [
          "id",
          "run_id",
          "ordinal",
          "name",
          "kind",
          "status",
          "depends_on",
          "output_ref",
          "retry_count",
          "started_at",
          "completed_at",
          "error_code"
        ],
        "title": "RunStepView",
        "type": "object"
      },
...
      "StepKind": {
        "enum": [
          "UNDERSTAND",
          "PLAN",
          "MODEL",
          "CAPABILITY",
          "VERIFY",
          "RESPOND"
        ],
        "title": "StepKind",
        "type": "string"
      },
      "StepStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "WAITING_APPROVAL",
          "COMPLETED",
          "FAILED",
          "SKIPPED",
          "CANCELLED"
        ],
        "title": "StepStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Enterprise Agent Runtime and Intelligence Workspace",
    "title": "Obsion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/evidence/{evidence_id}": {
      "get": {
        "operationId": "get_evidence_api_v1_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Evidence",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/artifacts": {
      "get": {
        "operationId": "list_artifacts_api_v1_runs__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactView"
                  },
                  "title": "Response List Artifacts Api V1 Runs  Run Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Artifacts",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/claims": {
      "get": {
        "operationId": "list_claims_api_v1_runs__run_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClaimView"
                  },
                  "title": "Response List Claims Api V1 Runs  Run Id  Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Claims",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/evidence": {
      "get": {
        "operationId": "list_evidence_api_v1_runs__run_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceView"
                  },
                  "title": "Response List Evidence Api V1 Runs  Run Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Evidence",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/steps": {
      "get": {
        "operationId": "list_steps_api_v1_runs__run_id__steps_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunStepView"
                  },
                  "title": "Response List Steps Api V1 Runs  Run Id  Steps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Steps",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}